requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
cloudinary==1.28.0
celery==5.4.0
coverage==6.2
cron-descriptor==1.4.3
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dj-database-url==0.5.0
django-celery-beat==2.6.0
django-timezone-field==7.0
Django==3.2.9
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
factory-boy==3.2.1
Faker==10.0.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.3.1
numpy==2.0.1
pandas==2.2.2
Pillow==6.2.2
psycopg2-binary==2.9.9
psycopg2==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.2.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
redis==5.0.7
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.6.0
